refactor: remove AbstractWriter and move its functionality to FileWriter and SystemMetadata
