+++ libs/output-mapping/src/SystemMetadata.php
@@ -5,13 +5,16 @@ declare(strict_types=1);
 namespace Keboola\OutputMapping;
 
 use Keboola\OutputMapping\Exception\OutputOperationException;
-use Keboola\OutputMapping\Writer\AbstractWriter;
-use Keboola\OutputMapping\Writer\TableWriter;
 
 class SystemMetadata
 {
     private array $systemMetadata;
 
+    public const SYSTEM_KEY_COMPONENT_ID = 'componentId';
+    public const SYSTEM_KEY_CONFIGURATION_ID = 'configurationId';
+    public const SYSTEM_KEY_CONFIGURATION_ROW_ID = 'configurationRowId';
+    public const SYSTEM_KEY_BRANCH_ID = 'branchId';
+    public const SYSTEM_KEY_RUN_ID = 'runId';
     public const SYSTEM_METADATA_PROVIDER = 'system';
     public const KBC_LAST_UPDATED_BY_BRANCH_ID = 'KBC.lastUpdatedBy.branch.id';
     public const KBC_LAST_UPDATED_BY_CONFIGURATION_ROW_ID = 'KBC.lastUpdatedBy.configurationRow.id';
@@ -25,7 +28,7 @@ class SystemMetadata
     public function __construct(array $systemMetadata)
     {
         $this->systemMetadata = $systemMetadata;
-        if (is_null($this->getSystemMetadata(AbstractWriter::SYSTEM_KEY_COMPONENT_ID))) {
+        if (is_null($this->getSystemMetadata(self::SYSTEM_KEY_COMPONENT_ID))) {
             throw new OutputOperationException('Component Id must be set');
         }
     }
@@ -38,11 +41,11 @@ class SystemMetadata
     public function getSystemTags(): array
     {
         $systemTags = [
-            AbstractWriter::SYSTEM_KEY_COMPONENT_ID,
-            AbstractWriter::SYSTEM_KEY_CONFIGURATION_ID,
-            AbstractWriter::SYSTEM_KEY_CONFIGURATION_ROW_ID,
-            AbstractWriter::SYSTEM_KEY_BRANCH_ID,
-            AbstractWriter::SYSTEM_KEY_RUN_ID,
+            self::SYSTEM_KEY_COMPONENT_ID,
+            self::SYSTEM_KEY_CONFIGURATION_ID,
+            self::SYSTEM_KEY_CONFIGURATION_ROW_ID,
+            self::SYSTEM_KEY_BRANCH_ID,
+            self::SYSTEM_KEY_RUN_ID,
         ];
 
         return array_filter($this->systemMetadata, function ($key) use ($systemTags) {
@@ -59,24 +62,24 @@ class SystemMetadata
     {
         $metadata[] = [
             'key' => self::KBC_CREATED_BY_COMPONENT_ID,
-            'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_COMPONENT_ID],
+            'value' => $this->systemMetadata[self::SYSTEM_KEY_COMPONENT_ID],
         ];
-        if (!empty($this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ID])) {
+        if (!empty($this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID])) {
             $metadata[] = [
                 'key' => self::KBC_CREATED_BY_CONFIGURATION_ID,
-                'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ID],
+                'value' => $this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID],
             ];
         }
-        if (!empty($this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ROW_ID])) {
+        if (!empty($this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID])) {
             $metadata[] = [
                 'key' => self::KBC_CREATED_BY_CONFIGURATION_ROW_ID,
-                'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ROW_ID],
+                'value' => $this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID],
             ];
         }
-        if (!empty($this->systemMetadata[AbstractWriter::SYSTEM_KEY_BRANCH_ID])) {
+        if (!empty($this->systemMetadata[self::SYSTEM_KEY_BRANCH_ID])) {
             $metadata[] = [
                 'key' => self::KBC_CREATED_BY_BRANCH_ID,
-                'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_BRANCH_ID],
+                'value' => $this->systemMetadata[self::SYSTEM_KEY_BRANCH_ID],
             ];
         }
         return $metadata;
@@ -86,24 +89,24 @@ class SystemMetadata
     {
         $metadata[] = [
             'key' => self::KBC_LAST_UPDATED_BY_COMPONENT_ID,
-            'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_COMPONENT_ID],
+            'value' => $this->systemMetadata[self::SYSTEM_KEY_COMPONENT_ID],
         ];
-        if (!empty($this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ID])) {
+        if (!empty($this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID])) {
             $metadata[] = [
                 'key' => self::KBC_LAST_UPDATED_BY_CONFIGURATION_ID,
-                'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ID],
+                'value' => $this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ID],
             ];
         }
-        if (!empty($this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ROW_ID])) {
+        if (!empty($this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID])) {
             $metadata[] = [
                 'key' => self::KBC_LAST_UPDATED_BY_CONFIGURATION_ROW_ID,
-                'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_CONFIGURATION_ROW_ID],
+                'value' => $this->systemMetadata[self::SYSTEM_KEY_CONFIGURATION_ROW_ID],
             ];
         }
-        if (!empty($this->systemMetadata[AbstractWriter::SYSTEM_KEY_BRANCH_ID])) {
+        if (!empty($this->systemMetadata[self::SYSTEM_KEY_BRANCH_ID])) {
             $metadata[] = [
                 'key' => self::KBC_LAST_UPDATED_BY_BRANCH_ID,
-                'value' => $this->systemMetadata[AbstractWriter::SYSTEM_KEY_BRANCH_ID],
+                'value' => $this->systemMetadata[self::SYSTEM_KEY_BRANCH_ID],
             ];
         }
         return $metadata;
